+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Change: Use HTTPS for link to WP SVN repository in bin script for configuring unit tests
 * Change: Note compatibility through WP 5.4+
 * Change: Update links to coffee2code.com to be HTTPS
+* Change: Unit tests: Remove unnecessary unregistering of hooks, and thus `tearDown()`
 
 ## 1.7.1 _(2019-11-12)_
 * Change: Note compatibility through WP 5.3+